requirements.txt
streamlit
biopython
pandas 